cpumask: Add cpumasks for big and little CPU clusters
Add cpu_lp_mask and cpu_perf_mask to represent the CPUs that belong to each cluster in a dual-cluster, heterogeneous system.

Note that this assumes that CONFIG_NR_CPUS is set to exatly the number of CPUs present, and that the number of CPUs per cluster is split evenly in half.

Original commit by kerneltoast