msm: kgsl: Wake GPU upon receiving an ioctl rather than upon touch input
Waking the GPU upon touch wastes power when the screen is being touched
in a way that does not induce animation or any actual need for GPU usage.
Instead of preemptively waking the GPU on touch input, wake it up upon
receiving a IOCTL_KGSL_GPU_COMMAND ioctl since it is a sign that the GPU
will soon be needed.

In my before testing gpu was boosting to 432MHz while scrolling static content.
With this patch it only goes up to 300MHz = more battery life!

Original commit by kerneltoast, adapted for s2 on 3.10 by nullby